refact: separate bip32 and bip39 code
